:bug: Economy.getMoney 오버플로우 현상 해결

+++ src/main/java/com/github/jaeukkang12/money/economy/Economy.java
@@ -26,7 +26,7 @@ public class Economy implements EconomyImpl {
         if(!moneyData.isExist(target.getUniqueId() + "")) {
             init(target);
         }
-        return moneyData.getInt(target.getUniqueId() + "");
+        return moneyData.getDouble(target.getUniqueId() + "");
     }
 
     public String getStringMoney(Player target) {
